# apps/web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cabinet CRM — façade de services",
    "description": "Façade HTTP localhost réutilisant le moteur Python du CRM. Toutes les routes /api/* exigent l'en-tête Authorization: Bearer <jeton>.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/patients/{patient_id}/clinical": {
      "get": {
        "tags": ["clinical"],
        "summary": "Patient Clinical",
        "operationId": "patient_clinical_api_patients__patient_id__clinical_get",
        "parameters": [
          {
            "name": "patient_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "title": "Patient Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ClinicalOut" } }
            }
          },
          "400": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            },
            "description": "Conflict"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/patients/{patient_id}/encaissements": {
      "get": {
        "tags": ["clinical"],
        "summary": "Patient Encaissements",
        "operationId": "patient_encaissements_api_patients__patient_id__encaissements_get",
        "parameters": [
          {
            "name": "patient_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "title": "Patient Id" }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "integer" }, { "type": "null" }],
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 0, "title": "Offset" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/EncaissementsOut" } }
            }
          },
          "400": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            },
            "description": "Conflict"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/patients/{patient_id}/creances": {
      "get": {
        "tags": ["clinical"],
        "summary": "Patient Creances",
        "operationId": "patient_creances_api_patients__patient_id__creances_get",
        "parameters": [
          {
            "name": "patient_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "title": "Patient Id" }
          },
          {
            "name": "include_notes",
            "in": "query",
            "required": false,
            "schema": { "type": "boolean", "default": false, "title": "Include Notes" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/CreanceOut" },
                  "title": "Response Patient Creances Api Patients  Patient Id  Creances Get"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            },
            "description": "Conflict"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/patients/{patient_id}/audit": {
      "get": {
        "tags": ["clinical"],
        "summary": "Patient Audit",
        "operationId": "patient_audit_api_patients__patient_id__audit_get",
        "parameters": [
          {
            "name": "patient_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "title": "Patient Id" }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 200, "title": "Limit" }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 0, "title": "Offset" }
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "enum": ["tous", "fiche", "plans", "actes", "paiements", "documents"],
              "type": "string",
              "default": "tous",
              "title": "Category"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/AuditOut" },
                  "title": "Response Patient Audit Api Patients  Patient Id  Audit Get"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            },
            "description": "Conflict"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/patients/{patient_id}/plans": {
      "post": {
        "tags": ["clinical"],
        "summary": "Plan Create",
        "operationId": "plan_create_api_patients__patient_id__plans_post",
        "parameters": [
          {
            "name": "patient_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "title": "Patient Id" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PlanIn" } } }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/PlanOut" } }
            }
          },
          "400": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            },
            "description": "Conflict"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/plans/{plan_id}": {
      "put": {
        "tags": ["clinical"],
        "summary": "Plan Update",
        "operationId": "plan_update_api_plans__plan_id__put",
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "title": "Plan Id" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PlanIn" } } }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/PlanOut" } }
            }
          },
          "400": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            },
            "description": "Conflict"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["clinical"],
        "summary": "Plan Delete",
        "operationId": "plan_delete_api_plans__plan_id__delete",
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "title": "Plan Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/OkOut" } }
            }
          },
          "400": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            },
            "description": "Conflict"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuditOut": {
        "properties": {
          "ts": { "type": "string", "title": "Ts" },
          "action": { "type": "string", "title": "Action" },
          "detail": { "title": "Detail" }
        },
        "type": "object",
        "required": ["ts", "action"],
        "title": "AuditOut"
      },
...
      "ClinicalOut": {
        "properties": {
          "notes_en_attente": {
            "items": { "$ref": "#/components/schemas/PaiementOut" },
            "type": "array",
            "title": "Notes En Attente"
          },
          "isoles": {
            "items": { "$ref": "#/components/schemas/PrestationOut" },
            "type": "array",
            "title": "Isoles"
          },
          "plans": {
            "items": { "$ref": "#/components/schemas/PlanGroupOut" },
            "type": "array",
            "title": "Plans"
          },
          "total_a_regler": { "type": "number", "title": "Total A Regler" }
        },
        "type": "object",
        "required": ["notes_en_attente", "isoles", "plans", "total_a_regler"],
        "title": "ClinicalOut"
      },
      "CreanceOut": {
        "properties": {
          "nature": { "type": "string", "title": "Nature" },
          "source_id": { "type": "integer", "title": "Source Id" },
          "libelle": { "type": "string", "title": "Libelle" },
          "montant": { "type": "number", "title": "Montant" },
          "reste": { "type": "number", "title": "Reste" },
          "date": { "anyOf": [{ "type": "string" }, { "type": "null" }], "title": "Date" },
          "statut": { "type": "string", "title": "Statut" }
        },
        "type": "object",
        "required": ["nature", "source_id", "libelle", "montant", "reste", "statut"],
        "title": "CreanceOut"
      },
...
      "EncaissementsOut": {
        "properties": {
          "items": {
            "items": { "$ref": "#/components/schemas/EncaissementOut" },
            "type": "array",
            "title": "Items"
          },
          "total": { "type": "integer", "title": "Total" },
          "solde": { "$ref": "#/components/schemas/TotauxOut" }
        },
        "type": "object",
        "required": ["items", "total", "solde"],
        "title": "EncaissementsOut"
      },
...
      "ErrorResponse": {
        "properties": { "error": { "$ref": "#/components/schemas/ErrorBody" } },
        "type": "object",
        "required": ["error"],
        "title": "ErrorResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OkOut": {
        "properties": { "ok": { "type": "boolean", "title": "Ok", "default": true } },
        "type": "object",
        "title": "OkOut"
      },
...
    }
  }
}